+++ CloudShell.Abstractions/ResourceManager/ResourceTabGroupIds.cs
@@ -3,6 +3,7 @@ namespace CloudShell.Abstractions.ResourceManager;
 public static class ResourceTabGroupIds
 {
     public const string General = "general";
+    public const string Application = "application";
     public const string Networking = "networking";
     public const string Runtime = "runtime";
     public const string Storage = "storage";